Write "chat room" as "chatroom" and "chat box" as chatbox"
We have bedroom, bathroom, classroom and boardroom as one word.
Living room is spelled separately because of the present participle (+ing) in living.

With "chatbox" it's a bit more subjective.
We have chatbot as one word (but litter box as two).
For consistency with chatroom, I'll go with one word.