WIP: Add support for XEP-0156.
It is disabled by default until more testing has been done, set
use_xep_0156 to true for now.

There is also the very big problem that the `Strophe.Connection` is
created before the user has had the time to enter their JID, which is
why we currently hardcode the domain to be resolved (this MUST be fixed
before this can be merged).  I haven’t been able to figure out where to
move this Connection creation, which must be after the JID is known but
before the actual connection is attempted.

There is also a bit of code duplication in it, this will be fixed in a
future version of this commit.

Fixes #129.
Fixes #1109.