Lazily compute the consistent colors for nicknames
When calling `occupant.getColor()` I think a color should always be
returned, but this means making the function async, and thereby using
`until` in the templates.

This change also only generates the color on demand, i.e. lazily,
instead of eagerly.

The benefit is that we avoid generating colors that aren't being used
because the occupant isn't being shown anywhere in the UI.