Initial work on MUC private messages
We now store messages on the `MUCOccupant` type, which is the first time
we're storing messages on a non-chatbox. So to handle this, we created
a new mixin model, `ModelWithMessages`, which can be extended by any
model (not just chatboxes) and give that model messaging ability.

We also turn ColorAwareModel and ModelWithMessages into mixins.

Updates #698