Refactor chat status
Converse has the concept of `status` for the logged-in user and
contacts, which is a combination of the presence `type` in some
cases and the `show` value.

I was refactoring `constructPresence` to try and be more explicit and
clear about the `type` and `show` values that an outgoing presence
should have, and while doing so run into the issue that those values
were being used together as the `status`.

In the end this turned out to be a much bigger refactor to clean up this
whole mess which has been there from the very early days.