bugfix: dragresize resistance not applied
When you're resizing, there should be some resistance around the default
height/view and it should snap to those defaults when you're within a
threshold. This makes it easier to resize a chat back to its original
dimensions.

This commit fixes that feature again.

It also refactors dragresize to be a mixin function instead of modifying
the prototype.