omemo: Remove all sessions when generating new keys
After generating new keys, when one sends out an encrypted message, it
must be a PreKeyWhisperMessage which contains a preKey to be used to
initiate a new session, otherwise the message can't be decrypted.

So we remove all existing sessions in order to force new sessions to be
created and thereby new PreKeyWhisperMessages to be sent out.

Updates #3769