Updates #3802
Trying to fix the `not-well-formed` reconnect loop.

- Don't reconnect when we get CONNFAIL error with `not-well-formed`.
- Don't flush immediately after connection when not using BOSH

Not sure about the 2nd change. The `not-well-formed` error happens
because the XML sent is somehow not valid. This reconnect loop happens
sometimes when resuming from suspend. My suspicion is that something
from the previous session is sent which is invalid in the current
session, but I'm not sure what, so this is mostly a stab in the dark.