Fixes #3824
Currently there are two different instances of DayJS. One in
`@converse/headless`, imported in `src/headless/shared/api/public.js`
and another in the UI code, which comes in via the dynamic import of the
DayJS locale.

The bug was that the locale was being set for the UI-code DayJS, but it
should be set for the headless one.

The UI-code DayJS is part of the dynamic import of the locale, but is
otherwise not used.

It would of course be nice to have only one instance, but I'm not sure
how this can be achieved while keeping dynamic imports of DayJS
translations in the UI-code.